+++ cmd/build.go
@@ -57,9 +57,9 @@ Complete documentation at https://mlange-42.github.io/modo/`,
 	root.Flags().BoolP("report-missing", "M", false, "Report missing docstings and coverage")
 	root.Flags().BoolP("case-insensitive", "C", false, "Build for systems that are not case-sensitive regarding file names.\nAppends hyphen (-) to capitalized file names")
 	root.Flags().BoolP("strict", "S", false, "Strict mode. Errors instead of warnings")
-	root.Flags().BoolP("dry-run", "D", false, "Dry-run without any file output")
+	root.Flags().BoolP("dry-run", "D", false, "Dry-run without any file output. Disables post-processing scripts")
 	root.Flags().BoolP("bare", "B", false, "Don't run pre- and post-processing scripts")
-	root.Flags().BoolVarP(&watch, "watch", "W", false, "Re-run on changes of sources and documentation files.\nDisables all post-processing scripts after running them once")
+	root.Flags().BoolVarP(&watch, "watch", "W", false, "Re-run on changes of sources and documentation files.\nDisables post-processing scripts after running them once")
 	root.Flags().StringSliceP("templates", "T", []string{}, "Optional directories with templates for (partial) overwrite.\nSee folder assets/templates in the repository")
 
 	root.Flags().SortFlags = false
@@ -95,7 +95,7 @@ func runBuild(args *document.Config) error {
 		return err
 	}
 
-	if !args.Bare {
+	if !args.Bare && !args.DryRun {
 		if err := runPostBuildCommands(args); err != nil {
 			return err
 		}